@@ -34,16 +34,13 @@ dev = [
 # Ruff configuration
 ############################
 [tool.ruff]
-target-version = "py310"
-line-length = 88
+line-length = 100
+target-version = "py311"
+# Keep only essential errors + import order for now
+select = ["E", "F", "I"]
+ignore = []
+extend-exclude = ["migrations", "alembic", "venv", ".venv", "build", "dist"]
 
-extend-select = [
-    "E", "F", "B", "I", "ARG", "PL",
-]
-
-ignore = [
-    "E501",
-]
 
 [tool.ruff.isort]
 known-first-party = ["app"]
@@ -53,16 +50,29 @@ combine-as-imports = true
 # mypy configuration
 ############################
 [tool.mypy]
-python_version = "3.10"
-strict = true
+# Runtime target for this repo
+python_version = "3.11"
+
+# Sprint-friendly defaults (advisory signal, not blocking)
+strict = false
+ignore_missing_imports = true
 warn_unused_configs = true
 warn_return_any = true
 warn_unreachable = true
 warn_redundant_casts = true
 warn_unused_ignores = true
-plugins = ["pydantic.mypy"]
-mypy_path = ["./api"]
 
+# If your code lives in api/app, tell mypy where to start
+# (Adjust if your package layout differs.)
+files = "app"
+
+# Useful plugin for Pydantic models
+plugins = ["pydantic.mypy"]
+
+# Historical: was set to ["./api"]; not needed when pyproject is already in api/
+# mypy_path = ["./api"]
+
+# Third-party/import-heavy modules we don’t type-check deeply
 [[tool.mypy.overrides]]
 module = [
     "redis.*",
@@ -73,8 +83,3 @@ module = [
 ]
 ignore_missing_imports = true
 
-############################
-# Bandit configuration
-############################
-[tool.bandit]
-skips = ["B101"]
