@@ -5,8 +5,12 @@ Run these checks during FastAPI lifespan to catch configuration errors early.
 from __future__ import annotations
 
 import logging
+import os
+import sys
 from collections import defaultdict
+from typing import Any
 
+from app.core.config import settings
 from app.models.base import Base
 
 logger = logging.getLogger(__name__)
@@ -25,7 +29,10 @@ def check_duplicate_table_mappings() -> None:
     table_to_mappers: dict[str, list[str]] = defaultdict(list)
 
     for mapper in Base.registry.mappers:
-        table_name = mapper.persist_selectable.name
+        # persist_selectable can be a Table, Subquery, or other selectable
+        # For mapped classes, it's typically a Table with a .name attribute
+        selectable: Any = mapper.persist_selectable
+        table_name = getattr(selectable, "name", str(selectable))
         class_name = mapper.class_.__name__
         table_to_mappers[table_name].append(class_name)
 
@@ -54,6 +61,32 @@ def check_duplicate_table_mappings() -> None:
     )
 
 
+def check_jwt_production_config() -> None:
+    """
+    Ensure JWT issuer/audience are set in production.
+
+    Fails fast if ENV=production and either JWT_ISSUER or JWT_AUDIENCE missing.
+    """
+    env = os.getenv("ENV", "development").lower()
+
+    if env == "production":
+        if not settings.jwt_issuer:
+            logger.error("❌ JWT_ISSUER is required in production")
+            sys.exit(1)
+
+        if not settings.jwt_audience:
+            logger.error("❌ JWT_AUDIENCE is required in production")
+            sys.exit(1)
+
+        logger.info(
+            "✓ JWT production config validated (iss=%s, aud=%s)",
+            settings.jwt_issuer,
+            settings.jwt_audience,
+        )
+    else:
+        logger.info("✓ JWT config optional in %s environment", env)
+
+
 def run_all_startup_checks() -> None:
     """
     Execute all startup integrity checks.
@@ -61,4 +94,5 @@ def run_all_startup_checks() -> None:
     """
     logger.info("Running startup integrity checks...")
     check_duplicate_table_mappings()
+    check_jwt_production_config()
     logger.info("All startup checks passed")
