@@ -1,11 +1,11 @@
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 from sqlalchemy import Index, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from .base import Base
-
-from typing import TYPE_CHECKING, List
 
 if TYPE_CHECKING:
     from .ro import RepairOrder
