@@ -1,8 +1,9 @@
 from datetime import datetime
 
-from app.models.base import Base
 from sqlalchemy import DateTime, String, func
 from sqlalchemy.orm import Mapped, mapped_column
+
+from app.models.base import Base
 
 
 class User(Base):
