@@ -1,8 +1,10 @@
 # api/app/schemas/ro.py
 from __future__ import annotations
+
 from datetime import datetime
-from pydantic import BaseModel
 from typing import Optional
+
+from pydantic import BaseModel
 
 
 class ROStatusMeta(BaseModel):
