@@ -4,7 +4,7 @@ module.exports = {
   env: { browser: true, es2023: true, node: true },
   parser: "@typescript-eslint/parser",
   parserOptions: { ecmaVersion: "latest", sourceType: "module", ecmaFeatures: { jsx: true } },
-  plugins: ["@typescript-eslint", "react", "react-hooks"],
+  plugins: ["@typescript-eslint", "react", "react-hooks", "es5"],
   extends: [
     "eslint:recommended",
     "plugin:react/recommended",
@@ -14,6 +14,7 @@ module.exports = {
   settings: { react: { version: "detect" } },
   ignorePatterns: ["dist/", "node_modules/", ".vite-cache/"],
   rules: {
-    "react/react-in-jsx-scope": "off" // React 17+ / Vite
+    "react/react-in-jsx-scope": "off",
+    "es5/no-es6-methods": "off"
   }
 };
