@@ -37,7 +37,7 @@ export default defineConfig(({ mode }) => {
       setupFiles: ["./vitest.setup.ts"],
       coverage: {
         provider: "v8",
-        reporter: ["text", "lcov"],
+        reporter: ["text", "json-summary", "lcov"],
       },
     },
   };
